Made grand refactoring and fixes
